requirements.txt
openai
ddgs
python-dotenv
pydantic
atproto
